backend/requirements.txt
fastapi
uvicorn
psycopg
python-dotenv